+++ CHANGELOG.rst
@@ -2,6 +2,30 @@
 Changelog
 =========
 
+Version 1.6
+===========
+* Minor change in dependencies for compatibility.
+
+Version 1.5
+===========
+* fit results are no longer `BenchmarkObservation`, and instead are moved into the datasets.
+
+Version 1.4
+===========
+
+* Renames:
+
+  * AnalysisResult -> BenchmarkAnalysisResult
+  * RunResult -> BenchmarkRunResult
+
+* Adds BenchmarkObservation class, and modifies BenchmarkAnalysisResult so observations now accepts a list[BenchmarkObservation].
+* Adds BenchmarkObservationIdentifier class.
+* Rebases RandomizedBenchmarking benchmarks, QuantumVolume, GHZ and CLOPS to use the new Observation class.
+* Fixes serialization of some circuits.
+* Adds AVAILABLE_BENCHMARKS to map a benchmark name to its class in __init__.
+* Adds benchmarks and configurations to __init__ for public import.
+* Other fixes.
+
 Version 1.3
 ===========